Decompose the reducer
