+++ src/constants.js
@@ -1,10 +1,21 @@
-const Movie = {
-  title: `The Grand Budapest Hotel`,
-  genre: `Drama`,
-  date: 2014,
-  poster: `img/the-grand-budapest-hotel-poster.jpg`,
-  bgImage: `img/bg-the-grand-budapest-hotel.jpg`,
-  src: `https://download.blender.org/durian/trailer/sintel_trailer-480p.mp4`,
+const emptyMovie = {
+  id: `Is loading...`,
+  title: ``,
+  poster: ``,
+  previewImage: ``,
+  bgImage: ``,
+  backgroundColor: ``,
+  src: ``,
+  previewVideoLink: ``,
+  description: ``,
+  rating: ``,
+  votes: ``,
+  director: ``,
+  starring: ``,
+  runTime: ``,
+  genre: ``,
+  date: ``,
+  isFavorite: ``,
 };
 
 const MovieTabs = {
@@ -17,4 +28,4 @@ const MAX_SHOWN_MOVIES = 8;
 
 const API_TIMEOUT = 5000;
 
-export {Movie, MovieTabs, MAX_SHOWN_MOVIES, API_TIMEOUT};
+export {emptyMovie, MovieTabs, MAX_SHOWN_MOVIES, API_TIMEOUT};
