+++ src/store/reducers/name-space.js
@@ -1,0 +1,6 @@
+const NameSpace = {
+  DATA: `DATA`,
+  APP_STATE: `APP_STATE`,
+};
+
+export default NameSpace;
